+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #1.1.5
 * Fix for the date on glossary update
 * Add link to Consistency tool on terms
+* On right click on the term with glossary is appended in the text in translation
 
 #1.1.4 
 * Improvement: Show the date of the last update